+++ src/TwoDStruct.h
@@ -2,16 +2,16 @@
 #define __TWODSTRUCT_H__
 class TwoDStruct{
   private:
-    double scale;
-    double rot;
-    double tranX;
-    double tranY;
+    float scale;
+    float rot;
+    float tranX;
+    float tranY;
   public:
-    TwoDStruct(double s, double r, double tX, double tY);
+    TwoDStruct(float s, float r, float tX, float tY);
     ~TwoDStruct();
-    double getScale();
-    double getRotate();
-    double getTranX();
-    double getTranY();
+    float getScale();
+    float getRotate();
+    float getTranX();
+    float getTranY();
 };
 #endif
